Update version to 2025.4.0 and transition to CalVer scheme (#29)
* Update version to 2025.4.0 and transition to CalVer scheme

This commit updates the versioning scheme to CalVer (Year.Minor.Micro) and sets the new version to 2025.4.0. It includes significant breaking changes, such as the removal of the factory pattern and the introduction of explicit constructors for data models. The `TypedTabularDataModel` has been replaced with a typed view from `TabularDataModel`, enhancing type safety and usability. Additionally, new utility functions for tabular processing have been added, and the documentation has been updated to reflect these changes, including a comprehensive migration guide for users. Tests and examples have been revised to align with the new structure and functionality.

* Remove deprecated ResourceManagerProtocol and associated methods; update README to reflect changes. Refactor TextFileHelper to remove deprecated load/load_as_stream methods, replacing them with read/read_as_stream. Update tests to ensure compatibility with new method names and functionality.